refactor: split profile routes (u/actors/orgs)
