+++ apps/web/src/app/profile/[id]/page.tsx
@@ -1,1203 +1,61 @@
-'use client';
+import { findUserByIdentifierWithSelect } from '@babylon/api';
+import { users } from '@babylon/db';
+import { loadActorsData } from '@babylon/engine';
+import { extractUsername } from '@babylon/shared';
+import { notFound, redirect } from 'next/navigation';
 
-import type { ProfileInfo } from '@babylon/shared';
-import {
-  type Actor,
-  cn,
-  extractUsername,
-  type FeedPost,
-  getBannerImageUrl,
-  isUsername,
-  type Organization,
-  POST_TYPES,
-} from '@babylon/shared';
-import {
-  ArrowLeft,
-  Coins,
-  FileText,
-  MessageCircle,
-  Search,
-} from 'lucide-react';
-import Link from 'next/link';
-import { useParams, useRouter } from 'next/navigation';
-import {
-  useCallback,
-  useEffect,
-  useLayoutEffect,
-  useMemo,
-  useState,
-} from 'react';
-import { ArticleCard } from '@/components/articles/ArticleCard';
-import { FollowButton } from '@/components/interactions/FollowButton';
-import { ModerationMenu } from '@/components/moderation/ModerationMenu';
-import { SendPointsModal } from '@/components/points/SendPointsModal';
-import { PostCard } from '@/components/posts/PostCard';
-import { EditProfileModal } from '@/components/profile/EditProfileModal';
-import { FollowListModal } from '@/components/profile/FollowListModal';
-import { OnChainBadge } from '@/components/profile/OnChainBadge';
-import {
-  type ProfileReply,
-  ProfileReplyCard,
-} from '@/components/profile/ProfileReplyCard';
-import { ProfileWidget } from '@/components/profile/ProfileWidget';
-import { Avatar } from '@/components/shared/Avatar';
-import { EmptyState } from '@/components/shared/EmptyState';
-import { PageContainer } from '@/components/shared/PageContainer';
-import {
-  FeedSkeleton,
-  ProfileHeaderSkeleton,
-} from '@/components/shared/Skeleton';
-import { VerifiedBadge } from '@/components/shared/VerifiedBadge';
-import { TradesFeed } from '@/components/trades/TradesFeed';
-import { useAuth } from '@/hooks/useAuth';
-import { useErrorToasts } from '@/hooks/useErrorToasts';
-import { useGameStore } from '@/stores/gameStore';
+export const dynamic = 'force-dynamic';
 
-export default function ActorProfilePage() {
-  const params = useParams();
-  const router = useRouter();
-  const identifier = decodeURIComponent(params.id as string);
-  const isUsernameParam = isUsername(identifier);
-  const actorId = isUsernameParam ? extractUsername(identifier) : identifier;
-  const { user, authenticated, getAccessToken } = useAuth();
-  const [searchQuery, setSearchQuery] = useState('');
-  const [tab, setTab] = useState<'posts' | 'replies' | 'trades'>('posts');
-  const { allGames } = useGameStore();
-  const [optimisticFollowerCount, setOptimisticFollowerCount] = useState<
-    number | null
-  >(null);
-  const [editProfileModalOpen, setEditProfileModalOpen] = useState(false);
-  const [followListModal, setFollowListModal] = useState<{
-    isOpen: boolean;
-    type: 'followers' | 'following';
-  }>({ isOpen: false, type: 'followers' });
+function equalsLoose(a: string, b: string): boolean {
+  return a.toLowerCase() === b.toLowerCase();
+}
 
-  // Check if viewing own profile - compare with both actorId and identifier (for ID-based URLs)
-  const isOwnProfile =
-    authenticated &&
-    user &&
-    (user.id === actorId ||
-      user.id === decodeURIComponent(identifier) ||
-      user.username === actorId ||
-      (user.username &&
-        user.username.startsWith('@') &&
-        user.username.slice(1) === actorId) ||
-      (user.username &&
-        !user.username.startsWith('@') &&
-        user.username === actorId));
+export default async function LegacyProfileIdPage({
+  params,
+}: {
+  params: Promise<{ id: string }>;
+}) {
+  const { id } = await params;
+  const identifier = decodeURIComponent(id);
 
-  // Use useLayoutEffect to redirect BEFORE paint to prevent flash of old username
-  // This runs synchronously before the browser paints, preventing any visual flash
-  useLayoutEffect(() => {
-    if (authenticated && user?.username && !isUsernameParam) {
-      const decodedIdentifier = decodeURIComponent(identifier);
-      const viewingOwnId =
-        user.id === actorId ||
-        user.id === decodedIdentifier ||
-        user.id === identifier;
+  const user = await findUserByIdentifierWithSelect(identifier, {
+    id: users.id,
+    username: users.username,
+    isActor: users.isActor,
+  });
 
-      if (viewingOwnId && user.username) {
-        // Additional null check to prevent redirecting to /profile/undefined
-        const cleanUsername = user.username.startsWith('@')
-          ? user.username.slice(1)
-          : user.username;
-        // Only redirect if cleanUsername is valid and the current URL doesn't already match
-        if (
-          cleanUsername &&
-          identifier !== cleanUsername &&
-          decodedIdentifier !== cleanUsername &&
-          actorId !== cleanUsername
-        ) {
-          // Use router.replace for client-side navigation (preserves React state)
-          router.replace(`/profile/${cleanUsername}`);
-        }
-      }
+  if (user && user.isActor !== true) {
+    if (user.username) {
+      const handle = extractUsername(user.username);
+      redirect(`/u/${encodeURIComponent(handle)}`);
     }
-  }, [
-    authenticated,
-    user?.id,
-    user?.username,
-    actorId,
-    identifier,
-    isUsernameParam,
-    router,
-  ]);
-
-  // Enable error toast notifications
-  useErrorToasts();
-
-  // Load actor/user info
-  const [actorInfo, setActorInfo] = useState<ProfileInfo | null>(null);
-  const [loading, setLoading] = useState(true);
-  const [isCreatingDM, setIsCreatingDM] = useState(false);
-  const [sendPointsModalOpen, setSendPointsModalOpen] = useState(false);
-  const [apiPosts, setApiPosts] = useState<
-    Array<{
-      id: string;
-      content: string;
-      author: string;
-      authorId: string;
-      timestamp: string;
-      authorName?: string;
-      authorUsername?: string | null;
-      authorProfileImageUrl?: string | null;
-      likeCount?: number;
-      commentCount?: number;
-      shareCount?: number;
-      isLiked?: boolean;
-      isShared?: boolean;
-      // Repost metadata
-      isRepost?: boolean;
-      isQuote?: boolean;
-      quoteComment?: string | null;
-      originalPostId?: string | null;
-      originalPost?: {
-        id: string;
-        content: string;
-        authorId: string;
-        authorName: string;
-        authorUsername: string | null;
-        authorProfileImageUrl: string | null;
-        timestamp: string;
-      } | null;
-    }>
-  >([]);
-  const [loadingPosts, setLoadingPosts] = useState(false);
-  const [replies, setReplies] = useState<ProfileReply[]>([]);
-  const [loadingReplies, setLoadingReplies] = useState(false);
-
-  // Handle creating DM with user or messaging own agent
-  const handleMessageClick = async () => {
-    if (!authenticated || !actorInfo?.id || isCreatingDM || !user?.id) return;
-
-    setIsCreatingDM(true);
-
-    // Check if this is the user's own agent - redirect to team chat and select the agent
-    if (actorInfo.isAgent && actorInfo.managedBy === user.id) {
-      router.push(
-        `/agents/team?selectAgent=${encodeURIComponent(actorInfo.id)}`
-      );
-      setIsCreatingDM(false);
-      return;
-    }
-
-    // Generate deterministic chat ID (same format as backend)
-    // Sort IDs to ensure consistency
-    const sortedIds = [user.id, actorInfo.id].sort();
-    const chatId = `dm-${sortedIds.join('-')}`;
-
-    // Navigate directly to chat page
-    // Chat will be created in DB when first message is sent
-    router.push(`/chats?chat=${chatId}&newDM=${actorInfo.id}`);
-
-    setIsCreatingDM(false);
-  };
-
-  const loadActorInfo = useCallback(async () => {
-    setLoading(true);
-
-    const token = await getAccessToken();
-    const headers: HeadersInit = {
-      'Content-Type': 'application/json',
-    };
-    if (token) {
-      headers['Authorization'] = `Bearer ${token}`;
-    }
-
-    // First, always try to load as a user by ID
-    // This handles IDs like "testuser-53618432" or Privy IDs
-    const userResponse = await fetch(
-      `/api/users/${encodeURIComponent(actorId)}/profile`,
-      { headers }
-    ).catch((error: Error) => {
-      console.error('Error loading user by ID:', error);
-      return null;
-    });
-
-    if (userResponse?.ok) {
-      const userData = await userResponse.json();
-      const user = userData.user;
-
-      // Database has two types of records in User table:
-      // 1. Human users (isActor=false) - real people who signed up
-      // 2. NPC records (isActor=true) - actors stored in User table for game interactions
-      //
-      // Skip NPC records - let the actor lookup below handle them
-      // since it has proper static image fallbacks (/images/actors/{id}.jpg)
-      //
-      // Also skip "shell" records that have no meaningful profile data
-      // (created when an org/actor ID was used as authorId for posts)
-      const isNonActorUser =
-        user &&
-        !user.isActor &&
-        (user.displayName ||
-          user.username ||
-          user.profileImageUrl ||
-          user.walletAddress);
-
-      if (isNonActorUser) {
-        setActorInfo({
-          id: user.id,
-          name: user.displayName || user.username || 'User',
-          description: user.bio || '',
-          role: user.isAgent ? 'Agent' : 'User',
-          type: 'user' as const,
-          isUser: !user.isAgent,
-          isAgent: user.isAgent || false,
-          managedBy: user.managedBy || null,
-          username: user.username,
-          profileImageUrl: user.profileImageUrl,
-          coverImageUrl: user.coverImageUrl,
-          stats: user.stats,
-        });
-
-        // Redirect to username-based URL if username exists and we're not already on it
-        if (user.username && !isUsernameParam && !isOwnProfile) {
-          const cleanUsername = user.username.startsWith('@')
-            ? user.username.slice(1)
-            : user.username;
-          // Only redirect if we're not already on the target URL (case-insensitive comparison)
-          if (cleanUsername.toLowerCase() !== actorId.toLowerCase()) {
-            router.replace(`/profile/${cleanUsername}`);
-            return;
-          }
-        }
-
-        setLoading(false);
-        return;
-      }
-    }
-
-    // If it's a username (starts with @) or looks like a username, try username lookup
-    if (
-      isUsernameParam ||
-      (!actorId.startsWith('did:privy:') &&
-        actorId.length <= 42 &&
-        !actorId.includes('-'))
-    ) {
-      const usernameLookupResponse = await fetch(
-        `/api/users/by-username/${encodeURIComponent(actorId)}`,
-        { headers }
-      ).catch((error: Error) => {
-        console.error('Error loading user by username:', error);
-        return null;
-      });
-
-      if (usernameLookupResponse?.ok) {
-        const usernameData = await usernameLookupResponse.json();
-        const user = usernameData.user;
-
-        // Skip NPC records (isActor=true) - let actor lookup handle them
-        if (user && !user.isActor) {
-          setActorInfo({
-            id: user.id,
-            name: user.displayName || user.username || 'User',
-            description: user.bio || '',
-            role: user.isAgent ? 'Agent' : 'User',
-            type: 'user' as const,
-            isUser: !user.isAgent,
-            isAgent: user.isAgent || false,
-            managedBy: user.managedBy || null,
-            username: user.username,
-            profileImageUrl: user.profileImageUrl,
-            coverImageUrl: user.coverImageUrl,
-            stats: user.stats,
-          });
-
-          // Redirect to username-based URL if we're on ID-based URL
-          if (!isUsernameParam && user.username && !isOwnProfile) {
-            const cleanUsername = user.username.startsWith('@')
-              ? user.username.slice(1)
-              : user.username;
-            // Only redirect if we're not already on the target URL (case-insensitive comparison)
-            if (cleanUsername.toLowerCase() !== actorId.toLowerCase()) {
-              router.replace(`/profile/${cleanUsername}`);
-              return;
-            }
-          }
-
-          setLoading(false);
-          return;
-        }
-      }
-    }
-
-    // Try to load from API endpoint (uses optimized server-side loader)
-    let actorsDb: { actors?: Actor[]; organizations?: Organization[] } = {
-      actors: [],
-      organizations: [],
-    };
-    try {
-      const response = await fetch('/api/actors');
-      if (response.ok) {
-        actorsDb = await response.json();
-      } else {
-        console.error('Failed to load actors:', response.status);
-      }
-    } catch (error) {
-      console.error('Error fetching actors:', error);
-    }
-
-    // Find actor by id, username, or name (case-insensitive for username and name)
-    const actorIdLower = actorId.toLowerCase();
-    let actor = actorsDb.actors?.find((a) => a.id === actorId);
-    if (!actor) {
-      actor = actorsDb.actors?.find(
-        (a) =>
-          'username' in a &&
-          typeof a.username === 'string' &&
-          a.username.toLowerCase() === actorIdLower
-      );
-    }
-    if (!actor) {
-      actor = actorsDb.actors?.find(
-        (a) => a.name.toLowerCase() === actorIdLower
-      );
-    }
-    if (actor) {
-      // Find which game this actor belongs to (use actual actor.id, not URL parameter)
-      let gameId: string | null = null;
-      for (const game of allGames) {
-        const allActors = [
-          ...(game.setup?.mainActors || []),
-          ...(game.setup?.supportingActors || []),
-          ...(game.setup?.extras || []),
-        ];
-        if (allActors.some((a) => a.id === actor.id)) {
-          gameId = game.id;
-          break;
-        }
-      }
-
-      // Fetch actor stats from database
-      let stats = { followers: 0, following: 0, posts: 0 };
-      const statsResponse = await fetch(
-        `/api/actors/${encodeURIComponent(actor.id)}/stats`
-      ).catch((error: Error) => {
-        console.error('Failed to load actor stats:', error);
-        return null;
-      });
-
-      if (statsResponse?.ok) {
-        const statsData = await statsResponse.json();
-        if (statsData.stats) {
-          stats = {
-            followers: statsData.stats.followers || 0,
-            following: statsData.stats.following || 0,
-            posts: statsData.stats.posts || 0,
-          };
-        }
-      }
-
-      setActorInfo({
-        id: actor.id,
-        name: actor.name,
-        description: actor.description,
-        profileDescription: actor.profileDescription,
-        tier: actor.tier,
-        domain: actor.domain,
-        personality: actor.personality,
-        affiliations: actor.affiliations,
-        role: actor.role || actor.tier || 'Actor',
-        type: 'actor' as const,
-        game: gameId ? { id: gameId } : undefined,
-        username: ('username' in actor
-          ? (actor.username as string)
-          : actor.id) as string | undefined, // Use username if available, fallback to ID
-        // Explicitly set profile image URL with fallback to static actor image
-        // Use same defensive pattern as organization fallback below
-        profileImageUrl:
-          'profileImageUrl' in actor &&
-          typeof actor.profileImageUrl === 'string' &&
-          actor.profileImageUrl
-            ? actor.profileImageUrl
-            : `/images/actors/${actor.id}.jpg`,
-        stats,
-      });
-      setLoading(false);
-      return;
-    }
-
-    // Find organization (case-insensitive for name)
-    let org = actorsDb.organizations?.find((o) => o.id === actorId);
-    if (!org) {
-      org = actorsDb.organizations?.find(
-        (o) => o.name.toLowerCase() === actorIdLower
-      );
-    }
-    if (org) {
-      // Fetch organization stats from database (orgs are also stored as actors)
-      let stats = { followers: 0, following: 0, posts: 0 };
-      const statsResponse = await fetch(
-        `/api/actors/${encodeURIComponent(org.id)}/stats`
-      ).catch((error: Error) => {
-        console.error('Failed to load organization stats:', error);
-        return null;
-      });
-
-      if (statsResponse?.ok) {
-        const statsData = await statsResponse.json();
-        if (statsData.stats) {
-          stats = {
-            followers: statsData.stats.followers || 0,
-            following: statsData.stats.following || 0,
-            posts: statsData.stats.posts || 0,
-          };
-        }
-      }
-
-      // Check if org has a non-empty profileImageUrl property
-      const orgProfileImageUrl =
-        'profileImageUrl' in org &&
-        typeof org.profileImageUrl === 'string' &&
-        org.profileImageUrl
-          ? org.profileImageUrl
-          : `/images/organizations/${org.id}.jpg`;
-
-      setActorInfo({
-        id: org.id,
-        name: org.name,
-        description: org.description,
-        profileDescription: org.profileDescription,
-        type: 'organization' as const,
-        role: 'Organization',
-        profileImageUrl: orgProfileImageUrl,
-        stats,
-      });
-      setLoading(false);
-      return;
-    }
-
-    // Not found
-    setActorInfo(null);
-    setLoading(false);
-  }, [
-    actorId,
-    allGames,
-    isOwnProfile,
-    isUsernameParam,
-    router,
-    getAccessToken,
-  ]);
-
-  useEffect(() => {
-    loadActorInfo();
-  }, [loadActorInfo]);
-
-  // Listen for profile updates (when user follows/unfollows someone)
-  useEffect(() => {
-    const handleProfileUpdate = () => {
-      // Reset optimistic count to trigger refetch from server
-      setTimeout(() => {
-        setOptimisticFollowerCount(null);
-        // Refetch actor info to get updated counts
-        loadActorInfo();
-      }, 1000); // Small delay to allow backend cache invalidation
-    };
-
-    window.addEventListener('profile-updated', handleProfileUpdate);
-    return () =>
-      window.removeEventListener('profile-updated', handleProfileUpdate);
-  }, [loadActorInfo]);
-
-  // Reset optimistic count when actorInfo changes (server data arrived)
-  useEffect(() => {
-    if (actorInfo && optimisticFollowerCount !== null) {
-      // Wait a bit then reset to use server value
-      const timer = setTimeout(() => {
-        setOptimisticFollowerCount(null);
-      }, 2000);
-      return () => clearTimeout(timer);
-    }
-    return undefined;
-  }, [actorInfo, optimisticFollowerCount]);
-
-  useEffect(() => {
-    const loadPosts = async () => {
-      if (!actorId) return;
-
-      setLoadingPosts(true);
-      // If we have actorInfo with ID, use that; otherwise use actorId (could be username)
-      const searchId = actorInfo?.id || actorId;
-      // Fetch posts from API by authorId
-      const response = await fetch(
-        `/api/posts?actorId=${encodeURIComponent(searchId)}&limit=100`
-      );
-      if (response.ok) {
-        const data = await response.json();
-        if (data.posts && Array.isArray(data.posts)) {
-          setApiPosts(data.posts);
-        }
-      }
-      setLoadingPosts(false);
-    };
-
-    // Load posts when actorInfo is available (has the correct ID)
-    if (actorInfo?.id) {
-      loadPosts();
-    }
-  }, [actorId, actorInfo?.id]);
-
-  // Load replies when tab changes to 'replies'
-  useEffect(() => {
-    if (tab !== 'replies') return;
-    if (!actorInfo?.id) return;
-
-    const controller = new AbortController();
-
-    const loadReplies = async () => {
-      setLoadingReplies(true);
-      try {
-        const token = await getAccessToken();
-        const headers: HeadersInit = { 'Content-Type': 'application/json' };
-        if (token) {
-          headers['Authorization'] = `Bearer ${token}`;
-        }
-
-        const response = await fetch(
-          `/api/users/${encodeURIComponent(actorInfo.id)}/posts?type=replies`,
-          { headers, signal: controller.signal }
-        );
-        if (response.ok) {
-          const data = await response.json();
-          const items = data?.data?.items ?? data?.items ?? [];
-          setReplies(items);
-        }
-      } catch (error) {
-        if (error instanceof Error && error.name !== 'AbortError') {
-          console.error('Failed to fetch replies:', error);
-        }
-      } finally {
-        setLoadingReplies(false);
-      }
-    };
-
-    loadReplies();
-
-    return () => controller.abort();
-  }, [tab, actorInfo?.id, getAccessToken]);
-
-  // Get posts for this actor from all games
-  const gameStorePosts = useMemo(() => {
-    const posts: Array<{
-      post: FeedPost;
-      gameId: string;
-      gameName: string;
-      timestampMs: number;
-    }> = [];
-
-    allGames.forEach((game) => {
-      game.timeline?.forEach((day) => {
-        day.feedPosts?.forEach((post) => {
-          if (post.author === actorId) {
-            const postDate = new Date(post.timestamp);
-            posts.push({
-              post,
-              gameId: game.id,
-              gameName: game.id,
-              timestampMs: postDate.getTime(),
-            });
-          }
-        });
-      });
-    });
-
-    // Sort by timestamp (newest first)
-    return posts.sort((a, b) => b.timestampMs - a.timestampMs);
-  }, [allGames, actorId]);
-
-  // Combine API posts and game store posts, removing duplicates
-  const actorPosts = useMemo(() => {
-    const combined: Array<{
-      post: FeedPost;
-      gameId: string;
-      gameName: string;
-      timestampMs: number;
-    }> = [];
-
-    // Add API posts first
-    apiPosts.forEach((apiPost) => {
-      combined.push({
-        post: {
-          id: apiPost.id,
-          day: 0,
-          content: apiPost.content,
-          author: apiPost.authorId,
-          authorName: apiPost.authorName || actorInfo?.name || apiPost.authorId,
-          authorUsername: apiPost.authorUsername || actorInfo?.username || null,
-          authorProfileImageUrl:
-            apiPost.authorProfileImageUrl || actorInfo?.profileImageUrl || null,
-          timestamp: apiPost.timestamp,
-          type: POST_TYPES.POST, // User-generated posts
-          sentiment: 0, // Neutral sentiment for user posts
-          clueStrength: 0, // User posts don't have clue strength
-          pointsToward: null, // User posts don't hint at yes/no
-          likeCount: apiPost.likeCount,
-          commentCount: apiPost.commentCount,
-          shareCount: apiPost.shareCount,
-          isLiked: apiPost.isLiked,
-          isShared: apiPost.isShared,
-          // Repost metadata (pass through from API)
-          isRepost: apiPost.isRepost,
-          isQuote: apiPost.isQuote,
-          quoteComment: apiPost.quoteComment,
-          originalPostId: apiPost.originalPostId,
-          originalPost: apiPost.originalPost,
-        },
-        gameId: '',
-        gameName: '',
-        timestampMs: new Date(apiPost.timestamp).getTime(),
-      });
-    });
-
-    // Add game store posts that aren't already in API posts
-    const apiPostIds = new Set(apiPosts.map((p) => p.id));
-    gameStorePosts.forEach((gamePost) => {
-      if (!apiPostIds.has(gamePost.post.id)) {
-        combined.push({
-          ...gamePost,
-          post: {
-            ...gamePost.post,
-            authorProfileImageUrl:
-              gamePost.post.authorProfileImageUrl ||
-              actorInfo?.profileImageUrl ||
-              null,
-          },
-        });
-      }
-    });
-
-    // Sort by timestamp (newest first)
-    return combined.sort((a, b) => b.timestampMs - a.timestampMs);
-  }, [apiPosts, gameStorePosts, actorInfo]);
-
-  // Separate posts and replies
-  const originalPosts = useMemo(() => {
-    return actorPosts.filter((item) => !item.post.replyTo);
-  }, [actorPosts]);
-
-  const replyPosts = useMemo(() => {
-    return actorPosts.filter((item) => item.post.replyTo);
-  }, [actorPosts]);
-
-  // Filter by tab
-  const tabFilteredPosts = useMemo(() => {
-    return tab === 'posts' ? originalPosts : replyPosts;
-  }, [tab, originalPosts, replyPosts]);
-
-  // Filter by search query
-  const filteredPosts = useMemo(() => {
-    if (!searchQuery.trim()) return tabFilteredPosts;
-
-    const query = searchQuery.toLowerCase();
-    return tabFilteredPosts.filter((item) =>
-      item.post.content?.toLowerCase().includes(query)
-    );
-  }, [tabFilteredPosts, searchQuery]);
-
-  // Loading or actor not found
-  if (loading) {
-    // If we're redirecting, don't show loading state
-    if (isOwnProfile && user?.username && !isUsernameParam) {
-      return null;
-    }
-
-    return (
-      <PageContainer noPadding className="min-h-screen">
-        <div className="flex flex-1 overflow-hidden">
-          {/* Main Content */}
-          <div className="flex min-w-0 flex-1 flex-col overflow-hidden">
-            <div className="flex-1 overflow-y-auto">
-              <ProfileHeaderSkeleton />
-              <div className="mt-4 border-border/5 border-t">
-                <FeedSkeleton count={5} />
-              </div>
-            </div>
-          </div>
-
-          {/* Right: Widget placeholder */}
-          <div className="hidden w-96 flex-shrink-0 flex-col bg-sidebar p-4 xl:flex" />
-        </div>
-      </PageContainer>
-    );
+    redirect(`/u/id/${encodeURIComponent(user.id)}`);
   }
 
-  if (!actorInfo) {
-    return (
-      <PageContainer noPadding className="flex flex-col">
-        <div className="sticky top-0 z-10 bg-background">
-          <div className="flex items-center gap-4 px-4 py-3">
-            <Link
-              href="/feed"
-              className="rounded-full p-2 transition-colors hover:bg-muted/50"
-            >
-              <ArrowLeft className="h-5 w-5" />
-            </Link>
-            <h1 className="font-bold text-xl">Profile Not Found</h1>
-          </div>
-        </div>
-        <div className="flex flex-1 flex-col items-center justify-center gap-4">
-          <p className="text-muted-foreground">
-            User or Actor &quot;{actorId}&quot; not found
-          </p>
-          <Link
-            href="/feed"
-            className="rounded-lg bg-primary px-6 py-3 font-semibold text-primary-foreground transition-all hover:bg-primary/90"
-          >
-            Back to Feed
-          </Link>
-        </div>
-      </PageContainer>
-    );
+  const { actors, organizations } = loadActorsData();
+  const idLower = identifier.toLowerCase();
+
+  const org =
+    organizations?.find((o) => o.id === identifier) ||
+    organizations?.find((o) => equalsLoose(o.name, identifier)) ||
+    organizations?.find((o) => (o as { username?: string }).username && equalsLoose((o as { username: string }).username, identifier));
+
+  if (org) {
+    redirect(`/orgs/${encodeURIComponent(org.id)}`);
   }
 
-  return (
-    <PageContainer noPadding className="flex flex-col">
-      {/* Main layout - responsive with optional sidebar on xl screens */}
-      <div className="flex flex-1 overflow-hidden">
-        {/* Main content */}
-        <div className="flex min-w-0 flex-1 flex-col overflow-hidden">
-          {/* Header */}
-          <div className="sticky top-0 z-10 bg-background/95 backdrop-blur-sm">
-            <div className="flex items-center gap-4 px-4 py-3">
-              <Link
-                href="/feed"
-                className="rounded-full p-2 transition-colors hover:bg-muted/50"
-              >
-                <ArrowLeft className="h-5 w-5" />
-              </Link>
-              <div className="flex-1">
-                <h1 className="font-bold text-xl">{actorInfo.name}</h1>
-                <p className="text-muted-foreground text-sm">
-                  {actorInfo.stats?.posts || actorPosts.length} posts
-                </p>
-              </div>
-            </div>
-          </div>
+  const actor =
+    actors?.find((a) => a.id === identifier) ||
+    actors?.find((a) => (a as { username?: string }).username && equalsLoose((a as { username: string }).username, identifier)) ||
+    actors?.find((a) => a.name.toLowerCase() === idLower);
 
-          {/* Content area */}
-          <div className="flex-1 overflow-y-auto">
-            {/* Profile Header */}
-            <div className="border-border border-b">
-              {/* Cover Image */}
-              <div className="relative h-[200px] bg-muted">
-                {(() => {
-                  // Get banner URL using the utility function (supports CDN)
-                  const bannerUrl =
-                    actorInfo.isUser &&
-                    actorInfo.type === 'user' &&
-                    'coverImageUrl' in actorInfo
-                      ? (actorInfo.coverImageUrl as string)
-                      : getBannerImageUrl(
-                          null,
-                          actorInfo.id,
-                          actorInfo.type === 'organization'
-                            ? 'organization'
-                            : 'actor'
-                        );
+  if (actor) {
+    redirect(`/actors/${encodeURIComponent(actor.id)}`);
+  }
 
-                  return bannerUrl ? (
-                    <img
-                      src={bannerUrl}
-                      alt={`${actorInfo.name} banner`}
-                      className="h-full w-full object-cover"
-                      onError={(e) => {
-                        // Fallback to gradient if image not found
-                        e.currentTarget.style.display = 'none';
-                        e.currentTarget.nextElementSibling?.classList.remove(
-                          'hidden'
-                        );
-                      }}
-                    />
-                  ) : null;
-                })()}
-                <div
-                  className={cn(
-                    'pointer-events-none absolute inset-0 h-full w-full bg-gradient-to-br from-primary/20 to-primary/5',
-                    actorInfo.type === 'actor' ||
-                      actorInfo.type === 'organization'
-                      ? 'hidden'
-                      : ''
-                  )}
-                />
-              </div>
+  if (user && user.isActor === true) {
+    redirect(`/actors/${encodeURIComponent(user.id)}`);
+  }
 
-              {/* Profile Info Container */}
-              <div className="px-4 pb-4">
-                {/* Top Row: Avatar + Action Buttons */}
-                <div className="mb-4 flex items-start justify-between">
-                  {/* Profile Picture - Overlapping cover */}
-                  <div className="-mt-16 sm:-mt-20 relative">
-                    <div className="h-32 w-32 overflow-hidden rounded-full border-4 border-background bg-background sm:h-36 sm:w-36">
-                      <Avatar
-                        id={actorInfo.id}
-                        name={
-                          (actorInfo.name ?? actorInfo.username ?? '') as string
-                        }
-                        type={
-                          actorInfo.type === 'organization'
-                            ? 'business'
-                            : actorInfo.isUser || actorInfo.type === 'user'
-                              ? 'user'
-                              : (actorInfo.type as 'actor' | undefined)
-                        }
-                        src={actorInfo.profileImageUrl || undefined}
-                        size="lg"
-                        className="h-full w-full"
-                      />
-                    </div>
-                  </div>
-
-                  {/* Action Buttons */}
-                  <div className="flex items-center gap-2 pt-3">
-                    {authenticated && user && user.id !== actorInfo.id && (
-                      <>
-                        {/* Message button - for regular users, own agents (team chat), and other users' agents (DM) */}
-                        {((actorInfo.isUser && actorInfo.type === 'user') ||
-                          actorInfo.isAgent) && (
-                          <button
-                            onClick={handleMessageClick}
-                            disabled={isCreatingDM}
-                            className="rounded-full border border-border p-2 transition-colors hover:bg-muted/50 disabled:cursor-not-allowed disabled:opacity-50"
-                            title={
-                              actorInfo.isAgent &&
-                              actorInfo.managedBy === user.id
-                                ? 'Message in Agents'
-                                : 'Send message'
-                            }
-                          >
-                            <MessageCircle className="h-5 w-5" />
-                          </button>
-                        )}
-                        {/* Send points button - available for ALL users including agents and actors */}
-                        <button
-                          onClick={() => setSendPointsModalOpen(true)}
-                          className="rounded-full border border-border p-2 transition-colors hover:bg-muted/50"
-                          title="Send points"
-                        >
-                          <Coins className="h-5 w-5" />
-                        </button>
-                        <FollowButton
-                          userId={actorInfo.id}
-                          size="md"
-                          variant="button"
-                          onFollowerCountChange={(delta) => {
-                            // Optimistically update the follower count based on current displayed value
-                            setOptimisticFollowerCount((prev) => {
-                              const currentCount =
-                                prev !== null
-                                  ? prev
-                                  : actorInfo.stats?.followers || 0;
-                              return Math.max(0, currentCount + delta); // Never go negative
-                            });
-                          }}
-                        />
-                        {/* Report button - only for users, not own profile */}
-                        {actorInfo.isUser && actorInfo.type === 'user' && (
-                          <ModerationMenu
-                            targetUserId={actorInfo.id}
-                            targetUsername={actorInfo.username ?? undefined}
-                            targetDisplayName={actorInfo.name ?? undefined}
-                            targetProfileImageUrl={
-                              actorInfo.profileImageUrl ?? undefined
-                            }
-                            onActionComplete={() => {
-                              // Refresh profile data after report
-                              loadActorInfo();
-                            }}
-                          />
-                        )}
-                      </>
-                    )}
-                    {isOwnProfile && (
-                      <button
-                        type="button"
-                        onClick={() => setEditProfileModalOpen(true)}
-                        className="rounded-full border border-border px-4 py-2 font-bold transition-colors hover:bg-muted/50"
-                      >
-                        Edit profile
-                      </button>
-                    )}
-                  </div>
-                </div>
-
-                {/* Name and Handle */}
-                <div className="mb-3">
-                  <div className="mb-0.5 flex items-center gap-1">
-                    <h2 className="font-bold text-xl">
-                      {actorInfo.name ?? actorInfo.username ?? ''}
-                    </h2>
-                    {actorInfo.type === 'actor' && !actorInfo.isUser && (
-                      <VerifiedBadge size="md" />
-                    )}
-                    {actorInfo.type === 'user' && (
-                      <OnChainBadge
-                        isRegistered={actorInfo.onChainRegistered ?? false}
-                        nftTokenId={actorInfo.nftTokenId ?? null}
-                        size="md"
-                      />
-                    )}
-                  </div>
-                  {actorInfo.username && (
-                    <p className="text-[15px] text-muted-foreground">
-                      @{actorInfo.username}
-                    </p>
-                  )}
-                </div>
-
-                {/* Description/Bio */}
-                {(actorInfo.profileDescription || actorInfo.description) && (
-                  <p className="mb-3 whitespace-pre-wrap text-[15px] text-foreground">
-                    {actorInfo.profileDescription || actorInfo.description}
-                  </p>
-                )}
-
-                {/* Stats */}
-                <div className="flex gap-4 text-[15px]">
-                  <button
-                    onClick={() =>
-                      setFollowListModal({ isOpen: true, type: 'following' })
-                    }
-                    className="hover:underline"
-                  >
-                    <span className="font-bold text-foreground">
-                      {actorInfo.stats?.following || 0}
-                    </span>
-                    <span className="ml-1 text-muted-foreground">
-                      Following
-                    </span>
-                  </button>
-                  <button
-                    onClick={() =>
-                      setFollowListModal({ isOpen: true, type: 'followers' })
-                    }
-                    className="hover:underline"
-                  >
-                    <span className="font-bold text-foreground">
-                      {optimisticFollowerCount !== null
-                        ? optimisticFollowerCount
-                        : actorInfo.stats?.followers || 0}
-                    </span>
-                    <span className="ml-1 text-muted-foreground">
-                      Followers
-                    </span>
-                  </button>
-                </div>
-              </div>
-            </div>
-
-            {/* Tabs: Posts vs Replies vs Trades */}
-            <div className="sticky top-0 z-10 border-border border-b bg-background/95 backdrop-blur-sm">
-              <div className="flex h-14 items-center justify-between px-4">
-                {/* Tab Buttons */}
-                <div className="flex flex-1 items-center">
-                  <button
-                    onClick={() => setTab('posts')}
-                    className={cn(
-                      'relative h-full px-4 font-semibold transition-all duration-300 hover:bg-muted/30',
-                      tab === 'posts'
-                        ? 'text-foreground opacity-100'
-                        : 'text-foreground opacity-50'
-                    )}
-                  >
-                    Posts
-                  </button>
-                  <button
-                    onClick={() => setTab('replies')}
-                    className={cn(
-                      'relative h-full px-4 font-semibold transition-all duration-300 hover:bg-muted/30',
-                      tab === 'replies'
-                        ? 'text-foreground opacity-100'
-                        : 'text-foreground opacity-50'
-                    )}
-                  >
-                    Replies
-                  </button>
-                  <button
-                    onClick={() => setTab('trades')}
-                    className={cn(
-                      'relative h-full px-4 font-semibold transition-all duration-300 hover:bg-muted/30',
-                      tab === 'trades'
-                        ? 'text-foreground opacity-100'
-                        : 'text-foreground opacity-50'
-                    )}
-                  >
-                    Trades
-                  </button>
-                </div>
-
-                {/* Search Bar - Top Right (hide on trades tab) */}
-                {tab !== 'trades' && (
-                  <div className="relative w-64">
-                    <Search className="-translate-y-1/2 absolute top-1/2 left-3 h-4 w-4 text-muted-foreground" />
-                    <input
-                      type="text"
-                      value={searchQuery}
-                      onChange={(e) => setSearchQuery(e.target.value)}
-                      placeholder={`Search ${tab}...`}
-                      className="w-full rounded-full border-0 bg-muted py-2 pr-4 pl-10 text-sm focus:outline-none focus:ring-2 focus:ring-primary"
-                    />
-                  </div>
-                )}
-              </div>
-            </div>
-
-            {/* Posts/Replies/Trades */}
-            <div className="px-4">
-              {tab === 'trades' ? (
-                <TradesFeed userId={actorInfo?.id} />
-              ) : tab === 'replies' ? (
-                // Replies tab - show actual comment replies
-                loadingReplies ? (
-                  <div className="w-full">
-                    <FeedSkeleton count={5} />
-                  </div>
-                ) : replies.length === 0 ? (
-                  <EmptyState
-                    icon={MessageCircle}
-                    title={searchQuery ? 'No replies found' : 'No replies yet'}
-                    description={
-                      searchQuery
-                        ? 'Try adjusting your search terms'
-                        : 'Replies to other posts will appear here'
-                    }
-                  />
-                ) : (
-                  <div>
-                    {replies
-                      .filter(
-                        (reply) =>
-                          !searchQuery.trim() ||
-                          reply.content
-                            ?.toLowerCase()
-                            .includes(searchQuery.toLowerCase())
-                      )
-                      .map((reply) => (
-                        <ProfileReplyCard
-                          key={reply.id}
-                          reply={reply}
-                          authorId={actorInfo?.id || ''}
-                          authorName={
-                            actorInfo?.name || actorInfo?.username || ''
-                          }
-                          authorUsername={actorInfo?.username || null}
-                          authorProfileImageUrl={
-                            actorInfo?.profileImageUrl || null
-                          }
-                        />
-                      ))}
-                  </div>
-                )
-              ) : loadingPosts ? (
-                <div className="w-full">
-                  <FeedSkeleton count={5} />
-                </div>
-              ) : filteredPosts.length === 0 ? (
-                <EmptyState
-                  icon={FileText}
-                  title={searchQuery ? 'No posts found' : 'No posts yet'}
-                  description={
-                    searchQuery
-                      ? 'Try adjusting your search terms'
-                      : 'Posts will appear here once published'
-                  }
-                />
-              ) : (
-                <div className="space-y-0">
-                  {filteredPosts.map((item, i) => {
-                    const postData = {
-                      id: item.post.id,
-                      type: item.post.type,
-                      content: item.post.content,
-                      fullContent: item.post.fullContent || null,
-                      articleTitle: item.post.articleTitle || null,
-                      byline: item.post.byline || null,
-                      biasScore: item.post.biasScore ?? null,
-                      category: item.post.category || null,
-                      authorId: item.post.author,
-                      authorName: item.post.authorName,
-                      authorUsername: item.post.authorUsername || null,
-                      authorProfileImageUrl:
-                        item.post.authorProfileImageUrl || null,
-                      timestamp: item.post.timestamp,
-                      likeCount: item.post.likeCount,
-                      commentCount: item.post.commentCount,
-                      shareCount: item.post.shareCount,
-                      isLiked: item.post.isLiked,
-                      isShared: item.post.isShared,
-                      // Repost metadata
-                      isRepost: item.post.isRepost || false,
-                      isQuote: item.post.isQuote || false,
-                      quoteComment: item.post.quoteComment || null,
-                      originalPostId: item.post.originalPostId || null,
-                      originalPost: item.post.originalPost || null,
-                    };
-
-                    return postData.type && postData.type === 'article' ? (
-                      <ArticleCard
-                        key={`${item.post.id}-${i}`}
-                        post={postData}
-                      />
-                    ) : (
-                      <PostCard
-                        key={`${item.post.id}-${i}`}
-                        post={postData}
-                        showInteractions={true}
-                        showCommentInputBar={false}
-                        onCommentClick={() =>
-                          router.push(`/post/${item.post.id}`)
-                        }
-                      />
-                    );
-                  })}
-                </div>
-              )}
-            </div>
-          </div>
-        </div>
-
-        {/* Widget Sidebar - Show for all profiles (users, actors, organizations) */}
-        {actorInfo && (
-          <div className="hidden w-96 flex-shrink-0 flex-col overflow-y-auto bg-sidebar p-4 xl:flex">
-            <ProfileWidget userId={actorInfo.id} />
-          </div>
-        )}
-      </div>
-
-      {/* Send Points Modal - Available for all users, agents, and actors */}
-      {actorInfo && (
-        <SendPointsModal
-          isOpen={sendPointsModalOpen}
-          onClose={() => setSendPointsModalOpen(false)}
-          recipientId={actorInfo.id}
-          recipientName={actorInfo.name ?? actorInfo.username ?? ''}
-          recipientUsername={actorInfo.username}
-          onSuccess={() => {
-            // Refresh profile data after successful transfer
-            loadActorInfo();
-          }}
-        />
-      )}
-
-      {/* Follow List Modal */}
-      {actorInfo && (
-        <FollowListModal
-          isOpen={followListModal.isOpen}
-          onClose={() =>
-            setFollowListModal({ ...followListModal, isOpen: false })
-          }
-          userId={actorInfo.id}
-          type={followListModal.type}
-        />
-      )}
-
-      {/* Edit Profile Modal (own profile only) */}
-      {isOwnProfile && (
-        <EditProfileModal
-          isOpen={editProfileModalOpen}
-          onClose={() => setEditProfileModalOpen(false)}
-        />
-      )}
-    </PageContainer>
-  );
+  notFound();
 }
